x86/resctrl: Ensure __rmid_read() is serialized on AMD
In AMD PQoS Versions 1.0 and 2.0, IA32_QM_EVTSEL MSR is shared by all
processors in a QOS domain.  So there's a chance it can read a different
event when two processors are reading the counter concurrently.

Normally event counter reads are serialized by the rdtgroup_mutex, but
the soft RMID implementation introduces IPI broadcasts to flush the
counters on all CPUs in a domain.

Obtain a spinlock in __rmid_read() when using soft RMIDs on affected
processors.

Co-developed-by: Peter Newman <peternewman@google.com>
Signed-off-by: Peter Newman <peternewman@google.com>
Signed-off-by: Stephane Eranian <eranian@google.com>