x86/resctrl: Add resctrl_mbm_flush_cpu() to collect CPUs' MBM events
AMD implementations so far are only guaranteed to provide MBM event
counts for RMIDs which are currently assigned in CPUs' PQR_ASSOC MSRs.
Hardware can reallocate the counter resources for all other RMIDs' which
are not currently assigned to those which are, zeroing the event counts
of the unassigned RMIDs.

In practice, this makes it impossible to simultaneously calculate the
memory bandwidth speed of all RMIDs on a busy system where all RMIDs are
in use. Over a multiple-second measurement window, the RMID would need
to remain assigned in all of the L3 cache domains where it has been
assigned for the duration of the measurement, otherwise portions of the
final count will be zero. In general, it is not possible to bound the
number of RMIDs which will be assigned in an L3 domain over any interval
of time.

To provide reliable MBM counts on such systems, introduce soft RMIDs:
when the mount option is enabled, each CPU is permanently assigned a
hardware RMID whose event counts are flushed to the current soft RMID
during context switches which result in a change in soft RMID.

Implement resctrl_mbm_flush_cpu(), which collects a domain's current MBM
event counts into its current soft RMID. The delta for each CPU is
determined by tracking the previous event counts in per-CPU data.  The
software byte counts reside in the arch_mbm_state structures.

To help resctrl_mbm_flush_cpu() locate the current CPU's L3 domain, add
CPU hotplug handlers to stash an rdt_domain pointer when it is safe to
call get_domain_from_cpu().

To measure the cost of soft RMID switching relative to that of normal
RMID and CLOSID switching, context switch cycle counts were obtained by
the perf tool monitoring a version of pipe-test-1m.c[1] modified to
assign the child process to a different resctrl group to control which
resources are switched.  The cycle counts below are simply the total
cycle count for the test job divided by 2 million, the number of context
switches.

Host: AMD EPYC 7B12 64-Core Processor

(unpatched)
	Same resctrl group	7803
	RMID switch		8302	(+6.40%)

(patched)
	Same resctrl group	7752	(-0.65%)
	RMID switch		8286	(+6.19%)
	Soft RMID		7796	(-0.09%)
	CLOSID+Soft RMID	8309	(+6.49%)

Host: Intel(R) Xeon(R) Gold 6268L CPU @ 2.80GHz

(unpatched)
	Same resctrl group	5166
	RMID switch		5810	(+12.46%)

(patched)
	Same resctrl group	5186	( +0.37%)
	RMID switch		5838	(+13.00%)
	Soft RMID		5198	( +0.68%)
	CLOSID+Soft RMID	5882	(+13.86%)

These results show that reading event counts is much cheaper than a
PQR_ASSOC write, but adds to that cost when switching both RMID and
CLOSID.

[1] https://people.redhat.com/mingo/cfs-scheduler/tools/pipe-test-1m.c

Co-developed-by: Stephane Eranian <eranian@google.com>
Signed-off-by: Stephane Eranian <eranian@google.com>
Signed-off-by: Peter Newman <peternewman@google.com>