x86/resctrl: Create soft RMID version of __mon_event_count()
When soft RMIDs are enabled, __mon_event_count() only needs to report the
current byte count in memory and should not touch RMIDs.

Create a parallel version for the soft RMID configuration and update
__mon_event_count() to choose between it and the original depending on
whether the soft RMID static key is enabled.

Create resctrl_arch_-accessors for the soft counts in arch_mbm_state as
__mon_event_count_soft_rmid() takes a struct rmid_read parameter, which
will only be visible to the fs code once it is moved.

Add clarifications to the documentation of various functions indicating
where a hardware rmid is required.

Signed-off-by: Peter Newman <peternewman@google.com>