x86/resctrl: Use overflow handler to push soft monitor counts
When using soft monitors, __mon_event_count() only reads the current
software count and does not cause CPUs in the domain to flush. For
mbm_handle_overflow() to be effective, it needs to flush the domain CPUs
so that all of the CPUs' RMIDs are read.

This avoids excessive IPIs when a user reads all event counts in all
RMIDs, it is more efficient to let the overflow handler drive the
propagation of hardware counter values to memory.

Signed-off-by: Peter Newman <peternewman@google.com>